summarize/requirements.txt
sumy
tinysegmenter
Janome
numpy
gensim
mecab-python-windows
